Split the code for the source and payload factories
Create new files for the source and payload factories and their tests.
